refactor: TimeRangeDto 추가
DTO 에서 날짜 형식을 통일하기 위해, 상속으로 변경

[skip ci]
