+++ server/edusync/src/main/java/com/codestates/edusync/model/study/plancalendar/dto/CalendarResponseDto.java
@@ -1,5 +1,6 @@
 package com.codestates.edusync.model.study.plancalendar.dto;
 
+import com.codestates.edusync.model.common.dto.TimeRangeDto;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -12,18 +13,13 @@ import java.util.Map;
 @NoArgsConstructor
 @Getter
 @Setter
-public class CalendarResponseDto {
+public class CalendarResponseDto extends TimeRangeDto.Response {
     private Long id;
     private Long groupId;
 
     private String title;
     private String content;
     private Boolean allDay;
-
-    @JsonFormat(pattern = "yyyy-MM-dd HH:mm", timezone = "Asia/Seoul")
-    private Timestamp startTime;
-    @JsonFormat(pattern = "yyyy-MM-dd HH:mm", timezone = "Asia/Seoul")
-    private Timestamp endTime;
 
     private String description;
     private Boolean overlap;
